combine levels to one game function
